requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1
python-docx==1.1.0 